I played around with proxying to the nodejs fileserver to hide internal details. and I got it working except for one thing -- couldn't delete files uploaded in the same browser session -- so I fell back to using direct server URL instead of /upload-gfx route
At this point, I've added a data attribute to the button so that I can
easily trigger a deletion based on filename (ie. derp.jpg, not
filepath), and I've hidden the buttons completely from view..

What really needs to happen next is the files need to be loaded in the
order they were queued when a new browser session is opened. I'm
thinking about putting a queue inside the main socketIO server